crud_app/requirement.txt
Django==5.2.9
djangorestframework
djangorestframework-simplejwt
django-rest-passwordreset
Pillow

